Split the Response object into respective DTO's in the model package in data layer
