+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/data/network/model/AiredDto.kt
@@ -1,0 +1,15 @@
+package com.abhijith.animex.data.network.model
+
+import com.google.gson.annotations.SerializedName
+
+data class AiredDto(
+
+    @SerializedName("string")
+    val airedInfo: String? = null,
+
+    @SerializedName("from")
+    val from: String? = null,
+
+    @SerializedName("to")
+    val to: Any? = null
+)
