+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/data/network/model/TitlesDto.kt
@@ -1,0 +1,13 @@
+package com.abhijith.animex.data.network.model
+
+import com.google.gson.annotations.SerializedName
+
+
+data class TitlesDto(
+
+    @SerializedName("type")
+    val type: String? = null,
+
+    @SerializedName("title")
+    val title: String? = null
+)
